Move Socials to containers
